requirements.txt
# ==================== FLASK CORE ====================
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Werkzeug==3.0.1

# ==================== DATABASE ====================
SQLAlchemy==2.0.23

# ==================== ENVIRONMENT & CONFIG ====================
python-dotenv==1.0.0

# ==================== SECURITY ====================
PyJWT==2.8.0
bcrypt==4.1.2

# ==================== MACHINE LEARNING ====================
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2

# TensorFlow (Optional - uncomment jika pakai deep learning)
# tensorflow==2.15.0

# PyTorch (Optional - uncomment jika pakai PyTorch)
# torch==2.1.0
# torchvision==0.16.0

# ==================== DATA PROCESSING ====================
openpyxl==3.1.2          # Untuk read Excel files
xlrd==2.0.1              # Untuk Excel support
python-dateutil==2.8.2

# ==================== PDF GENERATION ====================
reportlab==4.0.7         # Untuk generate PDF
matplotlib==3.8.2        # Untuk charts di PDF

# ==================== GOOGLE GEMINI API ====================
google-generativeai==0.3.2

# ==================== REDIS (Optional - untuk caching) ====================
# redis==5.0.1
# Flask-Caching==2.1.0

# ==================== API & HTTP ====================
requests==2.31.0

# ==================== UTILITIES ====================
python-multipart==0.0.6   # Untuk file upload
Pillow==10.1.0            # Untuk image processing (jika ada upload foto)

# ==================== PRODUCTION SERVER ====================
gunicorn==21.2.0          # Production WSGI server (untuk deploy)

# ==================== DEVELOPMENT TOOLS (Optional) ====================
# Flask-DebugToolbar==0.14.1
# pytest==7.4.3
# pytest-flask==1.3.0

# Existing dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
reportlab==4.0.7
matplotlib==3.8.2

# PostgreSQL
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2  # For reading Excel files

# Database Migrations (Optional but recommended)
alembic==1.12.1

# Development tools (optional)
jupyter==1.0.0
ipython==8.18.1